requirements.txt
bidict==0.23.1
mysql-connector-python==9.0.0
numpy==2.1.2
pandas==2.2.3
Requests==2.32.3
tqdm==4.66.5
scikit-learn==1.5.2
statsmodels==0.14.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cbor==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
fonttools==4.54.1
idna==3.10
ijson==3.3.0
inscriptis==2.5.0
ipython==8.12.3
ir_datasets==0.5.9
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lxml==5.3.0
lz4==4.3.3
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
rpds-py==0.20.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sns==0.1
soupsieve==2.6
stack-data==0.6.3
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trec-car-tools==2.6
tzdata==2024.2
unlzw3==0.2.2
urllib3==2.2.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zlib-state==0.1.9
python-terrier==0.10.0
repro-eval==0.4.0